renaming to weights and removing pub since there is no need for public referencing
