+++ pallets/identity/src/lib.rs
@@ -9,18 +9,18 @@ mod tests;
 #[cfg(feature = "runtime-benchmarks")]
 mod benchmarking;
 
-pub mod default_weights;
+mod weights;
 
 pub use pallet::*;
-pub use default_weights::*;
+use weights::*;
 
 #[frame_support::pallet]
 pub mod pallet {
-    use super::*;    
-    use frame_system::pallet_prelude::*;    
+    use super::*;
+    use frame_system::pallet_prelude::*;
     use codec::alloc::collections::BTreeSet;
     use frame_support::{dispatch::DispatchResult, inherent::Vec, pallet_prelude::*};
-    
+
 
     #[pallet::config]
     pub trait Config: frame_system::Config {
